requirements.txt
# Core dependencies
Pillow>=10.0.0
numpy>=1.24.0
cryptography>=41.0.0

# Optional dependencies
scipy>=1.10.0
flask>=2.3.0
